// operation-widget/models/IOperationWidgetConfig.ts
export interface IOperationWidgetConfig {
  device?: {
    id: string;
    name?: string;
  };
  buttons?: IOperationButtonConfig[];
}

export interface IOperationButtonConfig {
  label: string;
  icon?: string;
  operationFragment: string;
  buttonClasses?: string;
  description: string;
  operationValue: string;
  showModal: boolean;
  modalText?: string;
  customOperation?: boolean;
}